requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pdfminer.six
pymupdf
torch==2.2.0
torchvision==0.17.0
transformers
langchain
langchain-community
pymongo
python-dotenv
faiss-cpu
pydantic
pytest==8.1.1
pytest-asyncio==0.23.7
python-multipart
scikit-learn
scipy==1.9.3
numpy
rouge
openai==0.28.0
tqdm==4.66.2